+++ rollup.utils.js
@@ -1,10 +1,10 @@
-const fs = require("fs");
-const path = require("path");
 const babel = require("@rollup/plugin-babel").default;
-const nodeResolve = require("@rollup/plugin-node-resolve").default;
-const fse = require("fs-extra");
 const { camelCase, upperFirst } = require("lodash");
 const copy = require("rollup-plugin-copy");
+const fs = require("fs");
+const fse = require("fs-extra");
+const nodeResolve = require("@rollup/plugin-node-resolve").default;
+const path = require("path");
 
 const REPO_ROOT_DIR = __dirname;
 
@@ -191,28 +191,11 @@ function magicExportsPlugin({ packageName, version }) {
         banner +
         "\n" +
         "'use strict';\n" +
-        "Object.defineProperty(exports, '__esModule', { value: true });\n\n";
+        "Object.defineProperty(exports, '__esModule', { value: true });\n";
 
       if (magicExports.values) {
-        let deprecationFunctions =
-          // eslint-disable-next-line no-template-curly-in-string
-          "const getDeprecatedMessage = (symbol, packageName) => `All \\`remix\\` exports are considered deprecated as of v1.3.3. Please import \\`${symbol}\\` from \\`${packageName}\\` instead. Run \\`npx @remix-run/dev@latest codemod replace-remix-magic-imports\\` to automatically migrate your code.`;\n" +
-          "const warn = (fn, message) => (...args) => {\n" +
-          "  console.warn(message);\n" +
-          "  return fn(...args);\n" +
-          "};\n\n";
-
-        esmContents +=
-          `import * as ${moduleName} from '${packageName}';\n` +
-          deprecationFunctions;
-        esmContents += magicExports.values
-          .map(
-            (symbol) =>
-              `/** @deprecated Import \`${symbol}\` from \`${packageName}\` instead. */\n` +
-              `const ${symbol} = warn(${moduleName}.${symbol}, getDeprecatedMessage('${symbol}', '${packageName}'));\n`
-          )
-          .join("\n");
-        esmContents += `\nexport { ${magicExports.values.join(", ")} };\n`;
+        let exportList = magicExports.values.join(", ");
+        esmContents += `export { ${exportList} } from '${packageName}';\n`;
 
         tsContents += `import * as ${moduleName} from '${packageName}';\n\n`;
         tsContents += magicExports.values
@@ -223,15 +206,12 @@ function magicExportsPlugin({ packageName, version }) {
           )
           .join("\n");
 
-        cjsContents +=
-          `var ${moduleName} = require('${packageName}');\n` +
-          deprecationFunctions;
+        cjsContents += `var ${moduleName} = require('${packageName}');\n`;
         cjsContents += magicExports.values
           .map(
             (symbol) =>
               `/** @deprecated Import \`${symbol}\` from \`${packageName}\` instead. */\n` +
-              `const ${symbol} = warn(${moduleName}.${symbol}, getDeprecatedMessage('${symbol}', '${packageName}'));\n` +
-              `exports.${symbol} = ${symbol};\n`
+              `exports.${symbol} = ${moduleName}.${symbol};\n`
           )
           .join("\n");
       }