+++ scripts/publish.js
@@ -57,6 +57,7 @@ async function run() {
     "react",
     "serve",
     "css-bundle",
+    "testing",
   ]) {
     publish(path.join(buildDir, "@remix-run", name), tag);
   }