+++ packages/remix-express/CHANGELOG.md
@@ -1,5 +1,19 @@
 # `@remix-run/express`
 
+## 1.8.1
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/node@1.8.1`
+
+## 1.8.0
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/node@1.8.0`
+
 ## 1.7.6
 
 ### Patch Changes